Fixes in several file
Correctly setting locale, replace long at int and other type changes

+++ Chapter 5/Exercise 10/exercise_10.cpp
@@ -6,10 +6,10 @@ int main() {
     short amount = 0;
     std::cin >> amount;
 
-    for (int i = 1; i <= amount; ++i) {
-        for (int j = 1; j <= amount - i; ++j)
+    for (short i = 1; i <= amount; ++i) {
+        for (short j = 1; j <= amount - i; ++j)
             std::cout << '.';
-        for (int j = amount - i + 1; j <= amount; ++j)
+        for (short j = amount - i + 1; j <= amount; ++j)
             std::cout << '*';
         std::cout << std::endl;
     }
